+++ Exercism/c/atbash-cipher/src/atbash_cipher.h
@@ -1,0 +1,8 @@
+#ifndef ATBASH_CIHPER_H
+#define ATBASH_CIHPER_H
+
+char* atbash_encode(char* input);
+
+char* atbash_decode(char* input);
+
+#endif